fix an ambiguous compare
This function calls `shift` twice in the same expression.

```perl
"eq" => sub { shift->issued_token eq shift };
```

So, there is two possibilities of evaluating the expression.

```perl
"eq" => sub {
    my $first = shift;
    my $second = shift;
    return $fisrt->issued_token eq $second;
};
```

```perl
"eq" => sub {
    my $first = shift;
    my $second = shift;
    return $second->issued_token eq $fisrt;
};
```

I read the perlop, but the evaluating order is not defined.
https://perldoc.perl.org/5.30.0/perlop.html#Auto-increment-and-Auto-decrement

So we should assume that
there is no guarantee for perl how to evaluate the expression.